+++ MyRectangle.cs
@@ -8,22 +8,22 @@ using System.Threading.Tasks;
 
 namespace OOP_Paint {
     public class MyRectangle : MyFigure {
-        public Int32 Width { set; get; }
-        public Int32 Height { set; get; }
+        public Single Width { set; get; }
+        public Single Height { set; get; }
 
 
 
-        public MyRectangle(Int32 _x1, Int32 _y1, Int32 _x2, Int32 _y2, Pen _pen) : base(_pen) {
+        public MyRectangle(Single _x1, Single _y1, Single _x2, Single _y2, Pen _pen) : base(_pen) {
             //???Повторяющийся код
             InitializeFigure(_x1, _y1, _x2, _y2);
 
         }
-        public MyRectangle(Int32 _x1, Int32 _y1, Int32 _x2, Int32 _y2, Color _color) : base(_color) {
+        public MyRectangle(Single _x1, Single _y1, Single _x2, Single _y2, Color _color) : base(_color) {
             //???Повторяющийся код
             InitializeFigure(_x1, _y1, _x2, _y2);
 
         }
-        public void InitializeFigure(Int32 _x1, Int32 _y1, Int32 _x2, Int32 _y2) {
+        public void InitializeFigure(Single _x1, Single _y1, Single _x2, Single _y2) {
             Location = FindLeftUpCornerCoord(_x1, _y1, _x2, _y2);
             Width = Math.Abs(_x1 - _x2);
             Height = Math.Abs(_y1 - _y2);
@@ -31,7 +31,7 @@ namespace OOP_Paint {
         }
 
 
-        public void Resize(Int32 _x1, Int32 _y1, Int32 _x2, Int32 _y2) {
+        public void Resize(Single _x1, Single _y1, Single _x2, Single _y2) {
             InitializeFigure(_x1, _y1, _x2, _y2);
 
         }
